requirements.txt
aiofiles==0.6.0
pydantic==1.8.2
playwright==1.13.1
asyncpraw==7.4.0
aiohttp==3.7.4.post0
Jinja2==3.0.1
setuptools==57.4.0
py-memoize==1.1.1
